Rename MinimalImgMetadata to ImageMetadata.
This is just a little less verbose and doesn't have an ambiguous
shortening of the word "image". Also, this model will eventually
describe our "complete" (insofar as that's possible) set of metadata, at
which point the "minimal" part will be kinda outdated. Might as well get
rid of it now.