Replace logo with Shandra's.
This is a logo Shandra used in the original BIL info site, which scales
better than the prevous one (looks better at different sizes), can be
made to work on nearly any plain background, and is available at a much
higher resolution. This is a temporary fix, but overall I think it works
much better than the previous one. I still need to check with Shandra to
see if she has the svg file though.