Add link to ods version.
There's now a html link to both the ods and xls template on the upload
page. For the most part, it seems like the ods version just works as
long as you have the pyexcel-ods3 dependency installed. Also, I added in
an ods test file (metadata_test.ods) that's basically the same as the
excel version.