Better differentiate between submit and create.
When we first made the metadata and collection creation pages, we used
the word submit (because the user is submiting a form), but submit has
taken on a different meanining in the ingest portal. Users "create"
collections and metadata first. Once they are created, they can be
"submitted" and validated.