Adapt celery tasks for use on production machine.
This still needs a lot of work, but simple tests of just creating and deleting
a collection's staging area seems to work correctly. It might be nice to
preserve the old method, so I temporarily left it in as a comment. This should
either be changed shortly or removed entirely. No dead code! :P