Unlock collection and metadata if validation fails
If the validation step for submission fails, we should unlock the
collection and associated metadata so that the user can fix it.

Also, added some comments. We should probably refactor the collection
detail view as it's getting a little long and convoluted in places,
especially the recently added validation steps.