Data path in Collection. Fix changing UUID problem
There wasn't really a point in maintaining ImageData as a separate
model. It was a very small model (basically just the data_path) with a
one-to-one relationship with Collection, so it made sense to just merge
the path directly into the Collection as a string.

Also, the problem where the UUID reported on the creation page being
different than what showed up in the model after it was created is fixed
now (for the most part). The solution uses django caches. I'm not sure
if this is an ideal solution or not, but it works for now. One small
problem I noticed is that you can still have the resulting UUID change
if you have multiple instances of the create collection page open, but I
think we can live with that.