Add submission and validation status to Collection
This commit also formats things nicely in the Collection list view. We
use FA icons for locked status and validation status. This was mostly
prepared for a demo, and, as such, the code is in need of some cleanup.

For starters, there's some copy-paste code that should be consolidated.

Secondly, we should revisit how we check status. Since, a success state
is frozen, we should probably check that first, since that won't change.
Checking the celery task id first is probably fine in the short term,
but those results will expire at some point, which means it won't always
work for projects that succeeded a long time in the past.

Thirdly, the code to check this stuff is only getting called on some of
the views, so it's possible that if you go straight from submitting to
the Collection list, it won't update properly.

And, lastly, all of this code to check status could probably be done a
little more efficiently.