Refactor. Add comments. Set default filter status.
Previously, when the filter was based off of the lock state, the default
was to show the "unlocked", which makes sense since those are presumably
the ones the user is still working on. Now, the filter is based off of
submission status, so I set it to show Collections that haven't yet been
submitted for similar reasons.