Use datalist to suggest funder IDs.
Instead of using a drop down list, this is just a text field that
suggests/autcompletes using a list of known funder IDs. Users can still
enter their own funder ID if they want. If the list we use ever gets too
long, we'll probably want to load the list with some ajax.

We should probably use this method with a few other fields too.